Add hello plugin
Following the tutorial this commit adds the hello plugin and a timer resource
in order to emitt messages only every two seconds.
