+++ src/main.rs
@@ -1,32 +1,11 @@
 use bevy::prelude::*;
+use hello_plugin::HelloPlugin;
+
+mod hello_plugin;
 
 fn main() {
     App::build()
-        .add_startup_system(add_people.system())
-        .add_system(greet_world.system())
-        .add_system(greet_people.system())
+        .add_plugins(DefaultPlugins)
+        .add_plugin(HelloPlugin)
         .run();
-}
-
-struct Person;
-
-struct Name(String);
-
-fn add_people(mut cmds: Commands) {
-    cmds.spawn()
-        .insert(Person)
-        .insert(Name("Krin Mijevski".to_string()));
-    cmds.spawn()
-        .insert(Person)
-        .insert(Name("Siri Sirovich".to_string()));
-}
-
-fn greet_people(query: Query<&Name, With<Person>>) {
-    for person in query.iter() {
-        println!("I greet you traveler, my name is {}!", person.0);
-    }
-}
-
-fn greet_world() {
-    println!("Hello World!");
 }
